% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipRead.R
\name{clip.read}
\alias{clip.read}
\title{A function to read data from the clipboard}
\usage{
clip.read(format = "character", header = TRUE)
}
\arguments{
\item{format}{The format of the output, defaults to character}

\item{data}{The data that should be copied to the clipboard}
}
\description{
This function will copy data from the clipboard
}
\examples{
clip.read(data, format=c("character","dataFrame"), header=TRUE)
}
