% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipWrite.R
\name{clipWrite}
\alias{clipWrite}
\title{A function to copy data to the clipboard}
\usage{
clipWrite(data, sep = "\\t", row.names = FALSE, col.names = TRUE)
}
\arguments{
\item{data}{The data that should be copied to the clipboard}

\item{sep}{The separator which should be used in the copied output, defaults to "\t"}

\item{row.names}{Decides if the output should keep row names or not, defaults to FALSE}

\item{col.names}{Decides if the output should keep column names or not, defaults to TRUE}
}
\description{
This function will copy data to the clipboard and strucutre it in a way that works for Excel
}
\examples{
clip.write(data)
}
