% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorToCharacter.R
\name{factorToCharacter}
\alias{factorToCharacter}
\title{A convertion function}
\usage{
factorToCharacter(data)
}
\arguments{
\item{data}{The data frame or factor vector that should be converted}
}
\description{
This function will change a data frame's columns from factor to character
}
\examples{
factorToCharacter(data)
}
