% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strClean.R
\name{strClean}
\alias{strClean}
\title{A string function to remove whitespace}
\usage{
strClean(str)
}
\arguments{
\item{str}{The string that should have it's whitespace removed}
}
\description{
This function will remove all whitespace before and after a string
}
\examples{
strClean(str)
}
