% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strTrim.R
\name{strTrim}
\alias{strTrim}
\title{A string function to remove whitespace}
\usage{
strTrim(str)
}
\arguments{
\item{str}{The string that should have it's whitespace removed}
}
\description{
This function will remove all whitespace before and after a string
}
\examples{
strTrim(str)
}
